pyproject.toml
[project]
name = "build-mcp-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
